pyproject.toml
[project]
name = "listing-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.44",
    "celery>=5.6.2",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=6.0.1",
    "django-ninja>=1.5.3",
    "django-ninja-jwt>=5.4.4",
    "django-storages>=1.14.6",
    "django-treebeard>=4.8.0",
    "faststream[cli,kafka]>=0.6.6",
    "opensearch-py[async]>=2.0.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "uvloop>=0.22.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.9",
    "mypy>=1.19.1",
    "ruff>=0.14.14",
]


[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"  
strict_optional = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"

[[tool.mypy.overrides]]
module = "ninja.*"
ignore_missing_imports = true
